# phase-2/specs/001-todo-fullstack-webapp/contracts/api-contract.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  description: RESTful API for the Todo web application backend
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the application
      responses:
        '200':
          description: Application is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2026-01-30T10:00:00Z"

  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      description: Retrieve all tasks associated with the specified user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose tasks to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new task for a user
      description: Create a new task associated with the specified user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to create the task for
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by its ID and user ID
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user who owns the task
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: The ID of the task to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a specific task
      description: Update an existing task with the specified ID
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user who owns the task
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a specific task
      description: Delete an existing task with the specified ID
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user who owns the task
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: The ID of the task to delete
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: Toggle the completion status of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user who owns the task
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: integer
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          description: Optional task description
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Task completion status
          example: false
        user_id:
          type: integer
          description: ID of the user who owns this task
          example: 1
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2026-01-30T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: "2026-01-30T10:00:00Z"

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          description: Optional task description
          example: "Milk, eggs, bread"

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Updated task title
          example: "Buy groceries"
        description:
          type: string
          description: Updated task description
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Updated task completion status
          example: true

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Item not found"